add functionality: refactoring files names of orden service
