+++ order-service/src/main/java/com/nttdata/bootcamp/model/Order.java
@@ -1,5 +1,6 @@
 package com.nttdata.bootcamp.model;
 
+import com.nttdata.bootcamp.dto.OrderStatus;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -7,24 +8,17 @@ import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-import java.time.LocalDate;
-
-@Document(collection = "orders")
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
+@Document(collection = "orders")
 public class Order {
   @Id
-  private String orderId;
-  private String userId;
-  private Integer eventoId;
-  private String tituloEvento;
-  private String ticketId;
-  private LocalDate fecha;
-  private double monto;
-  private Integer cantidad;
-  private String metodoPago;
-  private boolean isCompleted = false;
-  private String payment_intent;
+  private String id;
+  private Integer userId;
+  private Integer ticketId;
+  private Integer eventId;
+  private Double price;
+  private OrderStatus status;
 }
