+++ order-service/src/main/java/com/nttdata/bootcamp/repository/OrderRepository.java
@@ -2,9 +2,8 @@ package com.nttdata.bootcamp.repository;
 
 import com.nttdata.bootcamp.model.Order;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
-import reactor.core.publisher.Flux;
 
 public interface OrderRepository extends ReactiveCrudRepository<Order, String> {
-  Flux<Order> findAllByUserId(String userId);
-  Flux<Order> findAllByEventoId(Integer eventoId);
+//  Flux<Order> findAllByUserId(String userId);
+//  Flux<Order> findAllByEventoId(Integer eventoId);
 }
