+++ order-service/src/main/java/com/nttdata/bootcamp/service/OrderService.java
@@ -1,16 +1,15 @@
 package com.nttdata.bootcamp.service;
 
+import com.nttdata.bootcamp.dto.OrchestratorResponseDto;
+import com.nttdata.bootcamp.dto.OrderRequestDto;
+import com.nttdata.bootcamp.dto.OrderResponseDto;
 import com.nttdata.bootcamp.model.Order;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 public interface OrderService {
-  Mono<Order> save(Order order);
-  Mono<Void> confirmPayment(String orderId);
-  Mono<Order> get(String orderId);
-  Flux<Order> getAll();
-  Mono<Void> delete(String orderId);
+  Mono<Order> createOrder(OrderRequestDto orderRequestDto);
+  Flux<OrderResponseDto> getAllOrder();
 
-  Flux<Order> getOrdersByUserId(String userId);
-  Flux<Order> getOrdersByEventId(Integer eventId);
+  Mono<Void> updateOrder(OrchestratorResponseDto responseDto);
 }
