+++ order-service/src/main/java/com/nttdata/bootcamp/service/impl/OrderServiceImpl.java
@@ -1,61 +1,94 @@
 package com.nttdata.bootcamp.service.impl;
 
-import com.nttdata.bootcamp.exceptions.NotFoundException;
+import com.nttdata.bootcamp.dto.*;
 import com.nttdata.bootcamp.model.Order;
 import com.nttdata.bootcamp.repository.OrderRepository;
 import com.nttdata.bootcamp.service.OrderService;
-import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.reactive.function.client.WebClient;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
+import reactor.core.publisher.Sinks;
 
 @Service
-@RequiredArgsConstructor
 public class OrderServiceImpl implements OrderService {
 
-  private final OrderRepository repository;
+  @Autowired
+  private WebClient webClient;
+
+  @Autowired
+  private Sinks.Many<OrchestratorRequestDto> sink;
+
+  @Autowired
+  private OrderRepository repository;
+
+  private Double price;
 
   @Override
-  public Mono<Order> save(Order order) {
-    return repository.save(order);
+  public Mono<Order> createOrder(OrderRequestDto orderRequestDto) {
+    return webClient
+            .get()
+            .uri("/api/v1/tickets/" + orderRequestDto.getTicketId())
+            .retrieve()
+            .bodyToMono(TicketDto.class)
+            .doOnNext(ticketDto -> {
+              this.price = ticketDto.getPrecio();
+            })
+            .flatMap(a -> repository.save(dtoToEntity(orderRequestDto, this.price)))
+            .doOnNext(e -> orderRequestDto.setOrderId(e.getId()))
+            .doOnNext(e -> emitEvent(orderRequestDto, this.price));
   }
 
   @Override
-  public Mono<Void> confirmPayment(String orderId) {
-    //TODO: reducir la cantidad de tickets disponibles en el tickets service
-    return repository.findById(orderId)
-            .switchIfEmpty(Mono.error(new NotFoundException("La orden con id " + orderId + " no existe")))
-            .doOnNext(existOrder -> existOrder.setCompleted(true))
-            .flatMap(repository::save).then();
+  public Flux<OrderResponseDto> getAllOrder() {
+    return repository.findAll().map(this::entityToDto);
   }
 
   @Override
-  public Mono<Order> get(String orderId) {
-    return repository.findById(orderId)
-            .switchIfEmpty(Mono.error(new NotFoundException("La orden con id " + orderId + " no existe")));
+  public Mono<Void> updateOrder(OrchestratorResponseDto responseDto) {
+    System.out.println("Response: " + responseDto.getStatus());
+    return repository.findById(responseDto.getOrderId())
+            .doOnNext(p -> p.setStatus(responseDto.getStatus()))
+            .flatMap(repository::save)
+            .then();
   }
 
-  @Override
-  public Flux<Order> getAll() {
-    return repository.findAll();
+  private void emitEvent(OrderRequestDto orderRequestDTO, Double price) {
+    sink.tryEmitNext(getOrchestratorRequestDto(orderRequestDTO, price));
   }
 
-  @Override
-  public Mono<Void> delete(String orderId) {
-    return repository.findById(orderId)
-            .switchIfEmpty(Mono.error(new NotFoundException("La orden con id " + orderId + " no existe")))
-            .flatMap(repository::delete);
+  private Order dtoToEntity(final OrderRequestDto dto, Double price) {
+    Order purchaseOrder = new Order();
+    purchaseOrder.setId(dto.getOrderId());
+    purchaseOrder.setTicketId(dto.getTicketId());
+    purchaseOrder.setUserId(dto.getUserId());
+    purchaseOrder.setStatus(OrderStatus.ORDER_CREATED);
+    purchaseOrder.setPrice(price);
+
+    purchaseOrder.setEventId(dto.getEventId());
+    return purchaseOrder;
   }
 
-  @Override
-  public Flux<Order> getOrdersByUserId(String userId) {
-    return repository.findAllByUserId(userId)
-            .switchIfEmpty(Mono.error(new NotFoundException("Las ordenes para el usuario " + userId + " no existen")));
+  public OrchestratorRequestDto getOrchestratorRequestDto(OrderRequestDto orderRequestDTO, Double price) {
+    OrchestratorRequestDto requestDTO = new OrchestratorRequestDto();
+    requestDTO.setUserId(orderRequestDTO.getUserId());
+    requestDTO.setAmount(price);
+    requestDTO.setOrderId(orderRequestDTO.getOrderId());
+    requestDTO.setTicketId(orderRequestDTO.getTicketId());
+    return requestDTO;
   }
 
-  @Override
-  public Flux<Order> getOrdersByEventId(Integer eventId) {
-    return repository.findAllByEventoId(eventId)
-            .switchIfEmpty(Mono.error(new NotFoundException("Las ordenes para el evento " + eventId + " no existen")));
+  private OrderResponseDto entityToDto(Order purchaseOrder) {
+    System.out.println("Purchase Order Status::"+purchaseOrder.getStatus());
+    OrderResponseDto dto = new OrderResponseDto();
+    dto.setOrderId(purchaseOrder.getId());
+    dto.setTicketId(purchaseOrder.getTicketId());
+    dto.setUserId(purchaseOrder.getUserId());
+    dto.setStatus(purchaseOrder.getStatus());
+    dto.setAmount(purchaseOrder.getPrice());
+
+    dto.setEventId(purchaseOrder.getEventId());
+    return dto;
   }
 }
